Alter command execution to make it easier to add new bridges.
