+++ relays/substrate/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 
 [dependencies]
+anyhow = "1.0"
 async-std = "1.9.0"
 async-trait = "0.1.42"
 codec = { package = "parity-scale-codec", version = "2.0.0" }
