+++ relays/substrate/src/cli.rs
@@ -17,11 +17,11 @@
 //! Deal with CLI args of substrate-to-substrate relay.
 
 use bp_message_lane::LaneId;
-use frame_support::weights::Weight;
-use sp_core::Bytes;
-use sp_finality_grandpa::SetId as GrandpaAuthoritiesSetId;
+use codec::{Encode, Decode};
 use sp_runtime::app_crypto::Ss58Codec;
 use structopt::{clap::arg_enum, StructOpt};
+
+use crate::rialto_millau::cli as rialto_millau;
 
 /// Parse relay CLI args.
 pub fn parse_args() -> Command {
@@ -68,207 +68,132 @@ pub enum Command {
 	DeriveAccount(DeriveAccount),
 }
 
+impl Command {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::InitBridge(arg) => arg.run().await?,
+			Self::RelayHeaders(arg) => arg.run().await?,
+			Self::RelayMessages(arg) => arg.run().await?,
+			Self::SendMessage(arg) => arg.run().await?,
+			Self::EncodeCall(arg) => arg.run().await?,
+			Self::EncodeMessagePayload(arg) => arg.run().await?,
+			Self::EstimateFee(arg) => arg.run().await?,
+			Self::DeriveAccount(arg) => arg.run().await?,
+		})
+	}
+}
+
 /// Start headers relayer process.
 #[derive(StructOpt)]
 pub enum RelayHeaders {
-	/// Relay Millau headers to Rialto.
-	MillauToRialto {
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		rialto_sign: RialtoSigningParams,
-		#[structopt(flatten)]
-		prometheus_params: PrometheusParams,
-	},
-	/// Relay Rialto headers to Millau.
-	RialtoToMillau {
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		millau_sign: MillauSigningParams,
-		#[structopt(flatten)]
-		prometheus_params: PrometheusParams,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::RelayHeaders),
+}
+
+impl RelayHeaders {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// Start message relayer process.
 #[derive(StructOpt)]
 pub enum RelayMessages {
-	/// Serve given lane of Millau -> Rialto messages.
-	MillauToRialto {
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		millau_sign: MillauSigningParams,
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		rialto_sign: RialtoSigningParams,
-		#[structopt(flatten)]
-		prometheus_params: PrometheusParams,
-		/// Hex-encoded lane id that should be served by the relay. Defaults to `00000000`.
-		#[structopt(long, default_value = "00000000")]
-		lane: HexLaneId,
-	},
-	/// Serve given lane of Rialto -> Millau messages.
-	RialtoToMillau {
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		rialto_sign: RialtoSigningParams,
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		millau_sign: MillauSigningParams,
-		#[structopt(flatten)]
-		prometheus_params: PrometheusParams,
-		/// Hex-encoded lane id that should be served by the relay. Defaults to `00000000`.
-		#[structopt(long, default_value = "00000000")]
-		lane: HexLaneId,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::RelayMessages),
+}
+
+impl RelayMessages {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// Initialize bridge pallet.
 #[derive(StructOpt)]
 pub enum InitBridge {
-	/// Initialize Millau headers bridge in Rialto.
-	MillauToRialto {
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		rialto_sign: RialtoSigningParams,
-		#[structopt(flatten)]
-		millau_bridge_params: MillauBridgeInitializationParams,
-	},
-	/// Initialize Rialto headers bridge in Millau.
-	RialtoToMillau {
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		millau_sign: MillauSigningParams,
-		#[structopt(flatten)]
-		rialto_bridge_params: RialtoBridgeInitializationParams,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::InitBridge),
+}
+
+impl InitBridge {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// Send bridge message.
 #[derive(StructOpt)]
 pub enum SendMessage {
-	/// Submit message to given Millau -> Rialto lane.
-	MillauToRialto {
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		#[structopt(flatten)]
-		millau_sign: MillauSigningParams,
-		#[structopt(flatten)]
-		rialto_sign: RialtoSigningParams,
-		/// Hex-encoded lane id. Defaults to `00000000`.
-		#[structopt(long, default_value = "00000000")]
-		lane: HexLaneId,
-		/// Dispatch weight of the message. If not passed, determined automatically.
-		#[structopt(long)]
-		dispatch_weight: Option<ExplicitOrMaximal<Weight>>,
-		/// Delivery and dispatch fee in source chain base currency units. If not passed, determined automatically.
-		#[structopt(long)]
-		fee: Option<bp_millau::Balance>,
-		/// Message type.
-		#[structopt(subcommand)]
-		message: ToRialtoMessage,
-		/// The origin to use when dispatching the message on the target chain. Defaults to
-		/// `SourceAccount`.
-		#[structopt(long, possible_values = &Origins::variants(), default_value = "Source")]
-		origin: Origins,
-	},
-	/// Submit message to given Rialto -> Millau lane.
-	RialtoToMillau {
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		#[structopt(flatten)]
-		rialto_sign: RialtoSigningParams,
-		#[structopt(flatten)]
-		millau_sign: MillauSigningParams,
-		/// Hex-encoded lane id. Defaults to `00000000`.
-		#[structopt(long, default_value = "00000000")]
-		lane: HexLaneId,
-		/// Dispatch weight of the message. If not passed, determined automatically.
-		#[structopt(long)]
-		dispatch_weight: Option<ExplicitOrMaximal<Weight>>,
-		/// Delivery and dispatch fee in source chain base currency units. If not passed, determined automatically.
-		#[structopt(long)]
-		fee: Option<bp_rialto::Balance>,
-		/// Message type.
-		#[structopt(subcommand)]
-		message: ToMillauMessage,
-		/// The origin to use when dispatching the message on the target chain. Defaults to
-		/// `SourceAccount`.
-		#[structopt(long, possible_values = &Origins::variants(), default_value = "Source")]
-		origin: Origins,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::SendMessage),
+}
+
+impl SendMessage {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// A call to encode.
 #[derive(StructOpt)]
 pub enum EncodeCall {
-	/// Encode Rialto's Call.
-	Rialto {
-		#[structopt(flatten)]
-		call: ToRialtoMessage,
-	},
-	/// Encode Millau's Call.
-	Millau {
-		#[structopt(flatten)]
-		call: ToMillauMessage,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::EncodeCall),
+}
+
+impl EncodeCall {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// A `MessagePayload` to encode.
 #[derive(StructOpt)]
 pub enum EncodeMessagePayload {
-	/// Message Payload of Rialto to Millau call.
-	RialtoToMillau {
-		#[structopt(flatten)]
-		payload: RialtoToMillauMessagePayload,
-	},
-	/// Message Payload of Millau to Rialto call.
-	MillauToRialto {
-		#[structopt(flatten)]
-		payload: MillauToRialtoMessagePayload,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::EncodeMessagePayload),
+}
+
+impl EncodeMessagePayload {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// Estimate Delivery & Dispatch Fee command.
 #[derive(StructOpt)]
 pub enum EstimateFee {
-	/// Estimate fee of Rialto to Millau message.
-	RialtoToMillau {
-		#[structopt(flatten)]
-		rialto: RialtoConnectionParams,
-		/// Hex-encoded id of lane that will be delivering the message.
-		#[structopt(long)]
-		lane: HexLaneId,
-		/// Payload to send over the bridge.
-		#[structopt(flatten)]
-		payload: RialtoToMillauMessagePayload,
-	},
-	/// Estimate fee of Rialto to Millau message.
-	MillauToRialto {
-		#[structopt(flatten)]
-		millau: MillauConnectionParams,
-		/// Hex-encoded id of lane that will be delivering the message.
-		#[structopt(long)]
-		lane: HexLaneId,
-		/// Payload to send over the bridge.
-		#[structopt(flatten)]
-		payload: MillauToRialtoMessagePayload,
-	},
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::EstimateFee),
+}
+
+impl EstimateFee {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 /// Given a source chain `AccountId`, derive the corresponding `AccountId` for the target chain.
@@ -279,122 +204,17 @@ pub enum EstimateFee {
 /// since messages sent over the bridge will be able to spend these.
 #[derive(StructOpt)]
 pub enum DeriveAccount {
-	/// Given Rialto AccountId, display corresponding Millau AccountId.
-	RialtoToMillau { account: AccountId },
-	/// Given Millau AccountId, display corresponding Rialto AccountId.
-	MillauToRialto { account: AccountId },
+	#[structopt(flatten)]
+	RialtoMillau(rialto_millau::DeriveAccount),
 }
 
-/// MessagePayload that can be delivered to message lane pallet on Millau.
-#[derive(StructOpt, Debug)]
-pub enum MillauToRialtoMessagePayload {
-	/// Raw, SCALE-encoded `MessagePayload`.
-	Raw {
-		/// Hex-encoded SCALE data.
-		data: Bytes,
-	},
-	/// Construct message to send over the bridge.
-	Message {
-		/// Message details.
-		#[structopt(flatten)]
-		message: ToRialtoMessage,
-		/// SS58 encoded account that will send the payload (must have SS58Prefix = 42)
-		#[structopt(long)]
-		sender: AccountId,
-	},
-}
-
-/// MessagePayload that can be delivered to message lane pallet on Rialto.
-#[derive(StructOpt, Debug)]
-pub enum RialtoToMillauMessagePayload {
-	/// Raw, SCALE-encoded `MessagePayload`.
-	Raw {
-		/// Hex-encoded SCALE data.
-		data: Bytes,
-	},
-	/// Construct message to send over the bridge.
-	Message {
-		/// Message details.
-		#[structopt(flatten)]
-		message: ToMillauMessage,
-		/// SS58 encoded account that will send the payload (must have SS58Prefix = 42)
-		#[structopt(long)]
-		sender: AccountId,
-	},
-}
-
-/// All possible messages that may be delivered to the Rialto chain.
-#[derive(StructOpt, Debug)]
-pub enum ToRialtoMessage {
-	/// Raw bytes for the message
-	Raw {
-		/// Raw, SCALE-encoded message
-		data: Bytes,
-	},
-	/// Make an on-chain remark (comment).
-	Remark {
-		/// Remark size. If not passed, small UTF8-encoded string is generated by relay as remark.
-		#[structopt(long)]
-		remark_size: Option<ExplicitOrMaximal<usize>>,
-	},
-	/// Transfer the specified `amount` of native tokens to a particular `recipient`.
-	Transfer {
-		/// SS58 encoded account that will receive the transfer (must have SS58Prefix = 42)
-		#[structopt(long)]
-		recipient: AccountId,
-		/// Amount of target tokens to send in target chain base currency units.
-		#[structopt(long)]
-		amount: bp_rialto::Balance,
-	},
-	/// A call to the Millau Bridge Message Lane pallet to send a message over the bridge.
-	MillauSendMessage {
-		/// Hex-encoded lane id that should be served by the relay. Defaults to `00000000`.
-		#[structopt(long, default_value = "00000000")]
-		lane: HexLaneId,
-		/// Raw SCALE-encoded Message Payload to submit to the message lane pallet.
-		#[structopt(long)]
-		payload: Bytes,
-		/// Declared delivery and dispatch fee in base source-chain currency units.
-		#[structopt(long)]
-		fee: bp_rialto::Balance,
-	},
-}
-
-/// All possible messages that may be delivered to the Millau chain.
-#[derive(StructOpt, Debug)]
-pub enum ToMillauMessage {
-	/// Raw bytes for the message
-	Raw {
-		/// Raw, SCALE-encoded message
-		data: Bytes,
-	},
-	/// Make an on-chain remark (comment).
-	Remark {
-		/// Size of the remark. If not passed, small UTF8-encoded string is generated by relay as remark.
-		#[structopt(long)]
-		remark_size: Option<ExplicitOrMaximal<usize>>,
-	},
-	/// Transfer the specified `amount` of native tokens to a particular `recipient`.
-	Transfer {
-		/// SS58 encoded account that will receive the transfer (must have SS58Prefix = 42)
-		#[structopt(long)]
-		recipient: AccountId,
-		/// Amount of target tokens to send in target chain base currency units.
-		#[structopt(long)]
-		amount: bp_millau::Balance,
-	},
-	/// A call to the Rialto Bridge Message Lane pallet to send a message over the bridge.
-	RialtoSendMessage {
-		/// Hex-encoded lane id that should be served by the relay. Defaults to `00000000`.
-		#[structopt(long, default_value = "00000000")]
-		lane: HexLaneId,
-		/// Raw SCALE-encoded Message Payload to submit to the message lane pallet.
-		#[structopt(long)]
-		payload: Bytes,
-		/// Declared delivery and dispatch fee in base source-chain currency units.
-		#[structopt(long)]
-		fee: bp_millau::Balance,
-	},
+impl DeriveAccount {
+	/// Run the command.
+	pub async fn run(self) -> anyhow::Result<()> {
+		Ok(match self {
+			Self::RialtoMillau(arg) => arg.run().await?,
+		})
+	}
 }
 
 arg_enum! {
@@ -456,7 +276,7 @@ impl AccountId {
 
 /// Lane id.
 #[derive(Debug)]
-pub struct HexLaneId(LaneId);
+pub struct HexLaneId(pub LaneId);
 
 impl From<HexLaneId> for LaneId {
 	fn from(lane_id: HexLaneId) -> LaneId {
@@ -471,6 +291,31 @@ impl std::str::FromStr for HexLaneId {
 		let mut lane_id = LaneId::default();
 		hex::decode_to_slice(s, &mut lane_id)?;
 		Ok(HexLaneId(lane_id))
+	}
+}
+
+/// Nicer formatting for raw bytes vectors.
+#[derive(Encode, Decode)]
+pub struct HexBytes(pub Vec<u8>);
+
+impl std::str::FromStr for HexBytes {
+	type Err = hex::FromHexError;
+
+	fn from_str(s: &str) -> Result<Self, Self::Err> {
+		Ok(Self(hex::decode(s)?))
+	}
+}
+
+impl std::fmt::Debug for HexBytes {
+	fn fmt(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+		write!(fmt, "0x{}", hex::encode(&self.0))
+	}
+}
+
+impl HexBytes {
+	/// Encode given object and wrap into nicely formatted bytes.
+	pub fn encode<T: Encode>(t: &T) -> Self {
+		Self(t.encode())
 	}
 }
 
@@ -527,6 +372,9 @@ where
 	}
 }
 
+/// Create chain-specific set of configuration objects: connection parameters,
+/// signing parameters and bridge initialisation parameters.
+#[macro_export]
 macro_rules! declare_chain_options {
 	($chain:ident, $chain_prefix:ident) => {
 		paste::item! {
@@ -557,17 +405,14 @@ macro_rules! declare_chain_options {
 			pub struct [<$chain BridgeInitializationParams>] {
 				#[doc = "Hex-encoded " $chain " header to initialize bridge with. If not specified, genesis header is used."]
 				#[structopt(long)]
-				pub [<$chain_prefix _initial_header>]: Option<Bytes>,
+				pub [<$chain_prefix _initial_header>]: Option<sp_core::Bytes>,
 				#[doc = "Hex-encoded " $chain " GRANDPA authorities set to initialize bridge with. If not specified, set from genesis block is used."]
 				#[structopt(long)]
-				pub [<$chain_prefix _initial_authorities>]: Option<Bytes>,
+				pub [<$chain_prefix _initial_authorities>]: Option<sp_core::Bytes>,
 				#[doc = "Id of the " $chain " GRANDPA authorities set to initialize bridge with. If not specified, zero is used."]
 				#[structopt(long)]
-				pub [<$chain_prefix _initial_authorities_set_id>]: Option<GrandpaAuthoritiesSetId>,
+				pub [<$chain_prefix _initial_authorities_set_id>]: Option<sp_finality_grandpa::SetId>,
 			}
 		}
 	};
 }
-
-declare_chain_options!(Rialto, rialto);
-declare_chain_options!(Millau, millau);
