+++ relays/substrate/src/rialto_millau/millau_messages_to_rialto.rs
@@ -19,7 +19,7 @@
 use crate::messages_lane::{select_delivery_transaction_limits, SubstrateMessageLane, SubstrateMessageLaneToSubstrate};
 use crate::messages_source::SubstrateMessagesSource;
 use crate::messages_target::SubstrateMessagesTarget;
-use crate::{MillauClient, RialtoClient};
+use super::{MillauClient, RialtoClient};
 
 use async_trait::async_trait;
 use bp_message_lane::{LaneId, MessageNonce};
