+++ relays/substrate/src/rialto_millau/rialto_headers_to_millau.rs
@@ -14,36 +14,36 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity Bridges Common.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Millau-to-Rialto headers sync entrypoint.
+//! Rialto-to-Millau headers sync entrypoint.
 
 use crate::{
 	finality_pipeline::{SubstrateFinalitySyncPipeline, SubstrateFinalityToSubstrate},
-	MillauClient, RialtoClient,
 };
+use super::{MillauClient, RialtoClient};
 
 use async_trait::async_trait;
-use relay_millau_client::{Millau, SyncHeader as MillauSyncHeader};
-use relay_rialto_client::{Rialto, SigningParams as RialtoSigningParams};
+use relay_millau_client::{Millau, SigningParams as MillauSigningParams};
+use relay_rialto_client::{Rialto, SyncHeader as RialtoSyncHeader};
 use relay_substrate_client::{finality_source::Justification, Error as SubstrateError, TransactionSignScheme};
 use sp_core::Pair;
 
-/// Millau-to-Rialto finality sync pipeline.
-pub(crate) type MillauFinalityToRialto = SubstrateFinalityToSubstrate<Millau, Rialto, RialtoSigningParams>;
+/// Rialto-to-Millau finality sync pipeline.
+pub(crate) type RialtoFinalityToMillau = SubstrateFinalityToSubstrate<Rialto, Millau, MillauSigningParams>;
 
 #[async_trait]
-impl SubstrateFinalitySyncPipeline for MillauFinalityToRialto {
-	const BEST_FINALIZED_SOURCE_HEADER_ID_AT_TARGET: &'static str = bp_millau::BEST_FINALIZED_MILLAU_HEADER_METHOD;
+impl SubstrateFinalitySyncPipeline for RialtoFinalityToMillau {
+	const BEST_FINALIZED_SOURCE_HEADER_ID_AT_TARGET: &'static str = bp_rialto::BEST_FINALIZED_RIALTO_HEADER_METHOD;
 
-	type SignedTransaction = <Rialto as TransactionSignScheme>::SignedTransaction;
+	type SignedTransaction = <Millau as TransactionSignScheme>::SignedTransaction;
 
 	async fn make_submit_finality_proof_transaction(
 		&self,
-		header: MillauSyncHeader,
-		proof: Justification<bp_millau::Header>,
+		header: RialtoSyncHeader,
+		proof: Justification<bp_rialto::Header>,
 	) -> Result<Self::SignedTransaction, SubstrateError> {
 		let account_id = self.target_sign.signer.public().as_array_ref().clone().into();
 		let nonce = self.target_client.next_account_index(account_id).await?;
-		let call = rialto_runtime::FinalityBridgeMillauCall::submit_finality_proof(
+		let call = millau_runtime::FinalityBridgeRialtoCall::submit_finality_proof(
 			header.into_inner(),
 			proof.into_inner(),
 			(),
@@ -51,23 +51,23 @@ impl SubstrateFinalitySyncPipeline for MillauFinalityToRialto {
 		.into();
 
 		let genesis_hash = *self.target_client.genesis_hash();
-		let transaction = Rialto::sign_transaction(genesis_hash, &self.target_sign.signer, nonce, call);
+		let transaction = Millau::sign_transaction(genesis_hash, &self.target_sign.signer, nonce, call);
 
 		Ok(transaction)
 	}
 }
 
-/// Run Millau-to-Rialto finality sync.
+/// Run Rialto-to-Millau finality sync.
 pub async fn run(
-	millau_client: MillauClient,
 	rialto_client: RialtoClient,
-	rialto_sign: RialtoSigningParams,
+	millau_client: MillauClient,
+	millau_sign: MillauSigningParams,
 	metrics_params: Option<relay_utils::metrics::MetricsParams>,
 ) {
 	crate::finality_pipeline::run(
-		MillauFinalityToRialto::new(rialto_client.clone(), rialto_sign),
-		millau_client,
+		RialtoFinalityToMillau::new(millau_client.clone(), millau_sign),
 		rialto_client,
+		millau_client,
 		metrics_params,
 	)
 	.await;
